Add confirmation dialog before deleting

+++ ppmtool-react-client/src/actions/projectActions.js
@@ -40,9 +40,15 @@ export const getProject = (id, history) => async dispatch => {
 };
 
 export const deleteProject = id => async dispatch => {
-  await axios.delete(`/api/project/${id}`);
-  dispatch({
-    type: DELETE_PROJECT,
-    payload: id
-  });
+  if (
+    window.confirm(
+      'Are you sure this will delete the project and all the data related to it?'
+    )
+  ) {
+    await axios.delete(`/api/project/${id}`);
+    dispatch({
+      type: DELETE_PROJECT,
+      payload: id
+    });
+  }
 };
